requirements.txt
Flask==3.0.3
Werkzeug==3.0.6
pillow==10.2.0
pytz==2024.1

